recipe-app-api/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.9.3
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0